mobile reader: extract useReaderChapter
Lifts chapter fetch effect (network-first → SQLite cache → empty-state)
out of the screen. Owns chapter/loading/chapterError state + wordCountRef.
